requirements_google.txt
# Google API dependencies (compatible with google-adk and google-cloud-storage)
google-auth>=2.26.1,<3.0.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.157.0,<3.0.0
